backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "Local RAG Pipeline Visualizer — FastAPI backend"
requires-python = ">=3.11,<3.14"

dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "python-multipart==0.0.9",
    "sentence-transformers==3.0.1",
    "faiss-cpu==1.8.0",
    "numpy==1.26.4",
    "httpx==0.27.2",
    "pydantic==2.8.2",
    "PyMuPDF==1.24.9",
    "python-dotenv==1.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.2",
]
